fix: 마이크 사용중 아이콘 사라지지 않는 문제 해결
